Remove OIDC auth

+++ src/main/java/com/alex/project/entiies/User.java
@@ -14,7 +14,7 @@ public class User extends PanacheEntityBase {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(name = "username")
     @NotNull
@@ -35,7 +35,7 @@ public class User extends PanacheEntityBase {
     @Enumerated(EnumType.STRING)
     private Role role;
 
-    public User(int id, String username, String password, Role role) {
+    public User(Long id, String username, String password, Role role) {
         this.id = id;
         this.username = username;
         this.password = password;
@@ -45,11 +45,11 @@ public class User extends PanacheEntityBase {
     public User() {
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
